+++ src/main.js
@@ -1,12 +1,7 @@
-/**
- * SPDX-FileCopyrightText: 2018 John Molakvoæ <skjnldsv@protonmail.com>
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
 import { generateFilePath } from '@nextcloud/router'
 
 import Vue from 'vue'
-import App from './App'
+import App from './App.vue'
 
 // eslint-disable-next-line
 __webpack_public_path__ = generateFilePath(appName, '', 'js/')