+++ stylelint.config.js
@@ -1,5 +1,3 @@
-// SPDX-FileCopyrightText: Nextcloud contributors
-// SPDX-License-Identifier: AGPL-3.0-or-later
 const stylelintConfig = require('@nextcloud/stylelint-config')
 
 module.exports = stylelintConfig