+++ templates/App.php
@@ -1,0 +1,3 @@
+<?php
+script('duplicatefinder', 'duplicatefinder-main');
+?>