+++ templates/settings-admin.php
@@ -1,6 +1,0 @@
-<?php
-script('duplicatefinder', 'settings');
-style('duplicatefinder', 'style');
-?>
-<div id="duplicatefinder-settings">
-</div>