+++ webpack.config.js
@@ -1,5 +1,8 @@
-// SPDX-FileCopyrightText: André Théo LAURET <andrelauret@eclipse-technology.eu>
-// SPDX-License-Identifier: AGPL-3.0-or-later
-const webpackConfig = require('@nextcloud/webpack-vue-config')
+const webpackConfig = require('@nextcloud/webpack-vue-config');
 
-module.exports = webpackConfig
+webpackConfig.entry = {
+    'main': './src/main.js',
+    'settings': './src/settings.js'
+};
+
+module.exports = webpackConfig;